# swagger.yaml
openapi: 3.0.0
info:
  title: TilTap API
  description: |
    Backend API for the TilTap speech-to-text and translation platform.
    Supports Telegram bot webhooks, translation proxy, and health checks.
  version: 1.0.0
  contact:
    name: TilTap Team
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/translate:
    post:
      tags: [Translate]
      summary: Translate text
      description: Proxy to the translation module. Falls back to OpenAI if no module URL is configured.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslateRequest"
      responses:
        "200":
          description: Translation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslateResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Translation service error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/web/translate:
    post:
      tags: [Web]
      summary: Translate text
      description: Direct text translation using the configured translation service.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslateRequest"
      responses:
        "200":
          description: Translation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslateResponse"
...
components:
  schemas:
    TranslateRequest:
      type: object
      required: [text, targetLang]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 5000
          description: Source text to translate
          example: "Салам, бу сөздү которуп берчи"
        targetLang:
          type: string
          minLength: 2
          maxLength: 10
          description: Target language code
          example: "ru"
        sourceLang:
          type: string
          minLength: 2
          maxLength: 10
          description: Optional source language code
          example: "ky"
    TranslateResponse:
      type: object
      properties:
        translatedText:
          type: string
          example: "Привет, переведи это слово"
        targetLang:
          type: string
          example: "ru"
        sourceLang:
          type: string
          example: "ky"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        details:
          type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT